Fixed a couple of narrowing cast bugs in AudioFormatReaders

+++ modules/juce_audio_formats/codecs/juce_FlacAudioFormat.cpp
@@ -248,7 +248,7 @@ public:
             }
             else
             {
-                if (startSampleInFile >= (int) lengthInSamples)
+                if (startSampleInFile >= lengthInSamples)
                 {
                     samplesInReservoir = 0;
                 }
